pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "email-mcp"
version = "0.1.0"
description = "Email MCP server with hybrid search, labels, and security tiers."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Ken Jorgensen" }]
dependencies = [
  "mcp[cli]>=1.0.0",
  "imapclient>=3.0.0",
  "sqlmodel>=0.0.16",
  "platformdirs>=4.2.0",
  "pydantic-settings>=2.2.0",
  "keyring>=25.0.0",
  "html2text>=2024.2.26",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
vector = [
  "sentence-transformers>=2.7.0",
  "chromadb>=0.5.0",
]

[project.scripts]
email-mcp = "email_mcp.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
